requirements.txt.txt
# Core FastAPI + Uvicorn stack (from your current venv)
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.1
colorama==0.4.6
fastapi==0.121.3
h11==0.16.0
httptools==0.7.1
idna==3.11
pip==25.3
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.50.0
typing_extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1

# Needed for file uploads (FastAPI's UploadFile / File / Form)
python-multipart

# HTTP calls (if your agents call APIs / LLMs)
requests

# Likely needed for your multi-agent / LLM supervisor (if used)
langchain
langchain-community
openai
langchain-openai